+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## 2.1.0 -- not released yet
 
+* Updated for Minecraft Bedrock 1.20.41 (@minecraft/server 1.7.0-beta).
 * New properties:
   * `Block.prototype.getItemStack`
   * `Block.prototype.type` setter